Former-commit-id: 91580a14a7840ddb3dce9e18543d91c24c9f8b11

+++ intermine/objectstore/test/src/org/intermine/util/CacheMapTest.java
@@ -1,7 +1,7 @@
 package org.intermine.util;
 
 /*
- * Copyright (C) 2002-2012 FlyMine
+ * Copyright (C) 2002-2013 FlyMine
  *
  * This code may be freely distributed and modified under the
  * terms of the GNU Lesser General Public Licence.  This should
@@ -27,7 +27,7 @@ public class CacheMapTest extends TestCase
 
         assertTrue(cm.size() < 1500);
         assertTrue("Expected first two to be missing",!(cm.containsKey(new Integer(2)) && cm.containsKey(new Integer(1))));
-        assertTrue("Expected last two to be present", cm.containsKey(new Integer(298)) || cm.containsKey(new Integer(297)));
+        assertTrue("Expected last two to be present", cm.containsKey(new Integer(298)) || cm.containsKey(new Integer(299)));
     }
 
     public void test2() throws Exception {
